requirements.txt
bs4
google-generativeai
pandas
requests
sentencepiece
transformers
